move osweb tests to os-webview repo and setup playwright2
