+++ e2e_tests/ui/test_details_page.py
@@ -1,6 +1,6 @@
 import pytest
 
-from e2e_tests.e2e.ui.pages.home import HomeRex
+from e2e_tests.ui.pages.home import HomeRex
 
 from playwright.async_api import TimeoutError
 import sys
