+++ e2e_tests/ui/test_osweb_higher_education_page.py
@@ -1,5 +1,5 @@
 import pytest
-from e2e_tests.e2e.ui.pages.home import HomeRex
+from e2e_tests.ui.pages.home import HomeRex
 
 
 @pytest.mark.asyncio
