+++ e2e_tests/ui/test_osweb_homepage.py
@@ -1,6 +1,6 @@
 import pytest
 
-from e2e_tests.e2e.ui.pages.home import HomeRex
+from e2e_tests.ui.pages.home import HomeRex
 
 
 @pytest.mark.asyncio
